+++ fuel-tx/src/tests/mod.rs
@@ -1,5 +1,4 @@
 mod offset;
-mod prepared_init;
 mod valid_cases;
 
 #[cfg(feature = "serde")]